listarVentas

+++ src/main/java/cl/playground/disconorte/repository/VentaMiembroRepository.java
@@ -3,8 +3,7 @@ package cl.playground.disconorte.repository;
 import cl.playground.disconorte.entity.VentaMiembro;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import java.lang.Integer;
 
 @Repository
-public interface VentaMiembroRepository extends JpaRepository<VentaMiembro, Integer> {
+public interface VentaMiembroRepository extends JpaRepository<VentaMiembro, Long> {
 }
